# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/repository/remainingnumber/JpaAnnLeaGrantRemDataRepo.java
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/nts.uk.com.web.nittsu.bundles.js
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/nts.uk.com.web.nittsu.bundles.js.map
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/ui/jquery-ext/gridlist-jquery-ext.ts
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/ui/ko-ext/datepicker-ko-ext.ts
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.infra/src/main/java/nts/uk/ctx/bs/employee/infra/repository/employee/history/AffCompanyHistRepositoryImp.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pub/src/main/java/nts/uk/ctx/bs/employee/pub/employee/export/PersonEmpBasicInfoPub.java
#	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.pubimp/src/main/java/nts/uk/ctx/bs/employee/pubimp/employee/export/PersonEmpBasicInfoPubImpl.java
#	nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/view/cps/001/a/cps001.a.vm.ts

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/empinfo/basicinfo/add/AddSpLea11InfoCommandHandler.java
@@ -14,12 +14,13 @@ import nts.uk.shr.pereg.app.command.PeregAddCommandHandler;
 import nts.uk.shr.pereg.app.command.PeregAddCommandResult;
 
 @Stateless
-public class AddSpLea11InfoCommandHandler extends CommandHandlerWithResult<AddSpecialleave11informationCommand, PeregAddCommandResult>
-implements PeregAddCommandHandler<AddSpecialleave11informationCommand>{
+public class AddSpLea11InfoCommandHandler
+		extends CommandHandlerWithResult<AddSpecialleave11informationCommand, PeregAddCommandResult>
+		implements PeregAddCommandHandler<AddSpecialleave11informationCommand> {
 
-	@Inject 
+	@Inject
 	private SpLeaInfoCommandHandler addSpLeaInfoCommandHandler;
-	
+
 	@Override
 	public String targetCategoryCd() {
 		return "CS00049";
@@ -32,12 +33,13 @@ implements PeregAddCommandHandler<AddSpecialleave11informationCommand>{
 
 	@Override
 	protected PeregAddCommandResult handle(CommandHandlerContext<AddSpecialleave11informationCommand> context) {
-		
+
 		val command = context.getCommand();
-		
-String cid = AppContexts.user().companyId();
-		
-		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00049.value, command.getUseAtr().intValue(), command.getAppSet().intValue(), command.getGrantDate(),
+
+		String cid = AppContexts.user().companyId();
+
+		SpecialLeaveBasicInfo domain = new SpecialLeaveBasicInfo(cid, command.getSID(), SpecialLeaveCode.CS00049.value,
+				command.getUseAtr(), command.getAppSet(), command.getGrantDate(),
 				command.getGrantDays() != null ? command.getGrantDays().intValue() : null, command.getGrantTable());
 		return new PeregAddCommandResult(addSpLeaInfoCommandHandler.addHandler(domain));
 	}
